fix: #2294 Set context on `Create From Order` smart browse.
